refactor: extract park to community repo + decouple resume (Path A.4)
Plugin moved to Soul-Brews-Studio/maw-park v0.1.0 (sha256-pinned in
maw-plugin-registry).

Pre-flight discovered audit-methodology gaps in a3-scout (#640):

1. SDK leaks: park/impl.ts imported `tmux` AND `hostExec` from
   ../../../sdk (= internal src/sdk, NOT the community-facing
   packages/sdk). Same class as rename's tmux leak. Tracked in #855.

2. Cross-plugin coupling: resume/index.ts imported cmdResume from
   ../park/impl — a sibling-plugin import. "Leaf" was wrong by this
   measure; park was the central node of a 2-plugin cluster.

Resolution:
- maw-park ships only cmdPark + cmdParkLs, refactored to spawnSync
  (bg-pattern) + 13 tests covering resolvePark + timeAgo pure logic.
- resume plugin in maw-js stays in core for now, but cmdResume + a
  fallback list helper are now inlined into resume/impl.ts. The cross-
  plugin import is gone.

LOC delta in maw-js: park/* removed (-150), resume/impl.ts added
(+95 cmdResume + helpers), resume/index.ts updated (1 line). Net -54.

Closes part of #640.
