--- src/commands/plugins/park/plugin.json
@@ -1,13 +1,0 @@
-{
-  "name": "park",
-  "version": "1.0.0",
-  "entry": "./index.ts",
-  "sdk": "^1.0.0",
-  "description": "Park (pause) an agent or list parked agents.",
-  "author": "Soul-Brews-Studio",
-  "cli": {
-    "command": "park",
-    "help": "maw park [agent] | maw park ls — park an agent or list parked agents"
-  },
-  "weight": 50
-}
